nested_enum_utils

Attribute Macro enum_conversions

#[enum_conversions]

Derive macro that generates conversions between an enum and its variants and other types.

The macro can be used as follows:

use nested_enum_utils::enum_conversions;

#[enum_conversions()]
enum MyEnum {
    Variant1(u32),
    Variant2(String),
}

This will create From instances from each variant type to the enum and TryFrom instances from the enum to each variant type.

The macro also accepts a list of target types to generate conversions to:

use nested_enum_utils::enum_conversions;

#[enum_conversions(Outer)]
enum Inner {
    Variant1(u32),
    Variant2(String),
}

#[enum_conversions()]
enum Outer {
    Inner1(Inner),
    // other variants
}

This will, in addition, generate From instances from each variant type to the outer enum and TryFrom instances from the outer enum to each variant type. The conversion to the outer enum relies on conversions between the inner enum and the outer enum, which is provided by the enum_conversions attribute on the Outer enum.

Limitations:

enums must have unnamed single fields

field types must be distinct